git-svn-id: https://taskunifier.svn.sourceforge.net/svnroot/taskunifier/TaskUnifierGui@264 4cc7d151-2d3f-4153-82be-2c8b9e7ce797

+++ src/com/leclercb/taskunifier/gui/components/configuration/GeneralConfigurationPanel.java
@@ -63,7 +63,7 @@ public class GeneralConfigurationPanel extends ConfigurationPanel {
 					(SimpleDateFormat) this.getValue("TIME_FORMAT"));
 			Settings.setLongProperty(
 					"synchronizer.scheduler_sleep_time",
-					((Long) this.getValue("SCHEDULER_SLEEP_TIME")) * 1000l);
+					((Integer) this.getValue("SCHEDULER_SLEEP_TIME")) * 1000l);
 		}
 	}
